requirements.txt
openai>=1.0.0
tqdm>=4.65.0
python-dotenv>=1.0.0
rich>=13.0.0
pytest>=7.0.0
jinja2>=3.1.0
markdown>=3.4.0
weasyprint>=60.0 